refactor: revert back to single phase
