+++ src/main/kotlin/ch/kleis/lcaplugin/core/lang/value/ConstraintValue.kt
@@ -1,7 +1,6 @@
 package ch.kleis.lcaplugin.core.lang.value
 
 import arrow.optics.optics
-import ch.kleis.lcaplugin.core.lang.expression.ConstraintFlag
 
 @optics
 sealed interface ConstraintValue {
@@ -14,7 +13,6 @@ object NoneValue : ConstraintValue
 data class FromProcessRefValue(
     val name: String,
     val arguments: Map<String, QuantityValue>,
-    val flag: ConstraintFlag = ConstraintFlag.NONE,
 ) : ConstraintValue {
     companion object
 }
