+++ src/test/kotlin/ch/kleis/lcaplugin/core/lang/evaluator/HelperTest.kt
@@ -1,6 +1,5 @@
-package ch.kleis.lcaplugin.core.lang.evaluator.compiler
+package ch.kleis.lcaplugin.core.lang.evaluator
 
-import ch.kleis.lcaplugin.core.lang.evaluator.compiler.Helper
 import ch.kleis.lcaplugin.core.lang.expression.*
 import ch.kleis.lcaplugin.core.lang.fixture.ProductFixture
 import ch.kleis.lcaplugin.core.lang.fixture.QuantityFixture
